+++ model_sms_send_request.go
@@ -21,7 +21,7 @@ type SmsSendRequest struct {
 	To string `json:"to"`
 	// The text of this message.
 	Text string `json:"text"`
-	// [Sender ID](https://help.ycloud.com/en/articles/3080386) to be used.
+	// [Sender ID](https://helpdocs.ycloud.com/help-center/integrations/channels/global-sms/sms-features/sender-id) to be used.
 	SenderId *string `json:"senderId,omitempty"`
 	// This parameter is only required for Chinese mainland SMS messages. You must specify an approved signature such as `Brand`. It will be added to the beginning of SMS body and wrapped with `【】`, e.g. `【Brand】Your verification code is 123456`.
 	Signature *string `json:"signature,omitempty"`